+++ CHANGELOG.md
@@ -21,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --- -->
 
+## [0.3.2] - 2026-03-04
+
+### Added
+
+- Add filtering functionality to MainContainer for inventory items
+
+---
+
 ## [0.3.1] - 2026-03-04
 
 ### Added